+++ src/string/define.js
@@ -3,3 +3,4 @@ import * as methods from './methods';
 
 defineStatic('range', methods.range);
 defineInstance('capitalize', methods.capitalize);
+defineInstance('pad', methods.pad);