+++ src/string/methods.js
@@ -1,2 +1,3 @@
 export { default as range } from './range';
 export { default as capitalize } from './capitalize';
+export { default as pad } from './pad';